Bound registration template diagnostics
