+++ test/shared/email.test.ts
@@ -1,180 +1,348 @@
 import { expect } from "@std/expect";
 import { describe, it as test } from "@std/testing/bdd";
-import type { EmailEntry } from "#shared/email.ts";
-import { buildSvgTicketData, buildTicketAttachments } from "#shared/email.ts";
-import { describeWithEnv } from "#test-utils/db.ts";
-import { makeTestEntry as makeEntry } from "#test-utils/factories.ts";
+import { spy } from "@std/testing/mock";
+import { t } from "#i18n";
+import type { EmailConfig, EmailMessage } from "#shared/email.ts";
+import {
+  deliverRegistrationEmail,
+  EMAIL_PROVIDER_LABELS,
+  sendEmail,
+  sendTestEmail,
+} from "#shared/email.ts";
+import { validEmail } from "#test-utils/email.ts";
+import { useFetchStub } from "#test-utils/mocks.ts";
 
-/** One group per entry — the shape a package-less order renders as. */
-const soloGroups = (entries: EmailEntry[]) =>
-  entries.map((entry) => ({ entries: [entry] }));
-
-const decodeAttachmentContent = (attachment: { content: string }): string => {
-  const binary = atob(attachment.content);
-  const bytes = Uint8Array.from(binary, (c) => c.charCodeAt(0));
-  return new TextDecoder().decode(bytes);
+const testConfig: EmailConfig = {
+  apiKey: "re_test_key",
+  fromAddress: validEmail("tickets@example.com"),
+  provider: "resend",
 };
 
-describeWithEnv("buildSvgTicketData", { db: true }, () => {
-  test("maps entry fields to SvgTicketData", () => {
-    const data = buildSvgTicketData(
-      makeEntry(
-        { name: "Concert" },
-        { price_paid: "1500", quantity: 2, ticket_token: "tok123" },
+const minimalMsg: EmailMessage = {
+  html: "h",
+  subject: "s",
+  text: "t",
+  to: validEmail("a@b.com"),
+};
+
+const plainMsg: EmailMessage = {
+  html: "<p>Hi</p>",
+  subject: "Test",
+  text: "Hi",
+  to: validEmail("user@test.com"),
+};
+
+const sendWithProvider = (
+  provider: EmailConfig["provider"],
+  msg: EmailMessage = minimalMsg,
+) => sendEmail({ ...testConfig, provider }, msg);
+
+const sendEmailExpectingError = async (
+  config: EmailConfig,
+  msg: EmailMessage,
+  expectedStatus: number | undefined,
+  expectedLogSubstring: string,
+): Promise<void> => {
+  const errorSpy = spy(console, "error");
+  try {
+    const status = await sendEmail(config, msg);
+    if (expectedStatus === undefined) {
+      expect(status).toBeUndefined();
+    } else {
+      expect(status).toBe(expectedStatus);
+    }
+    const logs = errorSpy.calls.map((c) => c.args[0] as string);
+    expect(logs.join("\n")).not.toContain(msg.to);
+    expect(
+      logs.some(
+        (l) => l.includes("E_EMAIL_SEND") && l.includes(expectedLogSubstring),
       ),
-      "GBP",
+    ).toBe(true);
+  } finally {
+    errorSpy.restore();
+  }
+};
+
+describe("sendEmail", () => {
+  const fetch = useFetchStub();
+  const mailgunBasicAuth = `Basic ${btoa("api:re_test_key")}`;
+
+  const expectMailgunRequest = (expectedUrl: string) => {
+    expect(fetch.callCount()).toBe(1);
+    const [url] = fetch.getFetchArgs();
+    expect(url).toBe(expectedUrl);
+    expect(fetch.getFetchHeaders().Authorization).toBe(mailgunBasicAuth);
+    const body = fetch.getFetchFormBody();
+    expect(body.get("from")).toBe("tickets@example.com");
+    expect(body.get("to")).toBe("user@test.com");
+    return body;
+  };
+
+  test("sends via Resend with correct URL, headers, and body", async () => {
+    const msg: EmailMessage = {
+      html: "<p>Hi</p>",
+      replyTo: validEmail("reply@test.com"),
+      subject: "Test",
+      text: "Hi",
+      to: validEmail("user@test.com"),
+    };
+
+    const status = await sendEmail(testConfig, msg);
+
+    expect(status).toBe(200);
+    expect(fetch.callCount()).toBe(1);
+    const [url, init] = fetch.getFetchArgs();
+    expect(url).toBe("https://api.resend.com/emails");
+    expect(init?.method).toBe("POST");
+    expect(fetch.getFetchHeaders().Authorization).toBe("Bearer re_test_key");
+    expect(fetch.getFetchHeaders()["Content-Type"]).toBe("application/json");
+    const body = fetch.getFetchJsonBody();
+    expect(body.from).toBe("tickets@example.com");
+    expect(body.to).toEqual(["user@test.com"]);
+    expect(body.reply_to).toBe("reply@test.com");
+    expect(body.subject).toBe("Test");
+    expect(body.html).toBe("<p>Hi</p>");
+    expect(body.text).toBe("Hi");
+  });
+
+  test("sends via Postmark with correct URL, headers, and body", async () => {
+    await sendWithProvider("postmark", {
+      html: "<p>Hi</p>",
+      subject: "Test",
+      text: "Hi",
+      to: validEmail("user@test.com"),
+    });
+
+    const [url] = fetch.getFetchArgs();
+    expect(url).toBe("https://api.postmarkapp.com/email");
+    expect(fetch.getFetchHeaders()["X-Postmark-Server-Token"]).toBe(
+      "re_test_key",
     );
-    expect(data.listingName).toBe("Concert");
-    expect(data.quantity).toBe(2);
-    expect(data.pricePaid).toBe("1500");
-    expect(data.checkinUrl).toContain("/checkin/tok123");
+    const body = fetch.getFetchJsonBody();
+    expect(body.From).toBe("tickets@example.com");
+    expect(body.To).toBe("user@test.com");
+    expect(body.Subject).toBe("Test");
+    expect(body.HtmlBody).toBe("<p>Hi</p>");
+    expect(body.TextBody).toBe("Hi");
   });
 
-  test("includes attendee date for daily listings", () => {
-    const data = buildSvgTicketData(
-      makeEntry({}, { date: "2026-06-15" }),
-      "GBP",
+  test("sends via SendGrid with correct URL, headers, and body", async () => {
+    const msg: EmailMessage = {
+      html: "<p>Hi</p>",
+      replyTo: validEmail("reply@test.com"),
+      subject: "Test",
+      text: "Hi",
+      to: validEmail("user@test.com"),
+    };
+
+    await sendWithProvider("sendgrid", msg);
+
+    const [url] = fetch.getFetchArgs();
+    expect(url).toBe("https://api.sendgrid.com/v3/mail/send");
+    expect(fetch.getFetchHeaders().Authorization).toBe("Bearer re_test_key");
+    const body = fetch.getFetchJsonBody();
+    expect(body.personalizations).toEqual([
+      { to: [{ email: "user@test.com" }] },
+    ]);
+    expect(body.from).toEqual({ email: "tickets@example.com" });
+    expect(body.reply_to).toEqual({ email: "reply@test.com" });
+    expect(body.content[0]).toEqual({ type: "text/plain", value: "Hi" });
+    expect(body.content[1]).toEqual({
+      type: "text/html",
+      value: "<p>Hi</p>",
+    });
+  });
+
+  test("sends via SendGrid without reply_to when not provided", async () => {
+    await sendWithProvider("sendgrid", {
+      html: "<p>Hi</p>",
+      subject: "Test",
+      text: "Hi",
+      to: validEmail("user@test.com"),
+    });
+
+    expect(fetch.getFetchJsonBody().reply_to).toBeUndefined();
+  });
+
+  test("sends via Mailgun (US) with correct URL, headers, and FormData body", async () => {
+    await sendWithProvider("mailgun-us", {
+      ...plainMsg,
+      replyTo: validEmail("reply@test.com"),
+    });
+
+    expect(fetch.getFetchHeaders()).not.toHaveProperty("Content-Type");
+    const body = expectMailgunRequest(
+      "https://api.mailgun.net/v3/example.com/messages",
     );
-    expect(data.attendeeDate).toBe("2026-06-15");
+    expect(body.get("subject")).toBe("Test");
+    expect(body.get("html")).toBe("<p>Hi</p>");
+    expect(body.get("text")).toBe("Hi");
+    expect(body.get("h:Reply-To")).toBe("reply@test.com");
   });
 
-  test("includes listing date and location from listing", () => {
-    const data = buildSvgTicketData(
-      makeEntry({ date: "2026-07-01T19:00:00Z", location: "Town Hall" }),
-      "GBP",
+  test("sends via Mailgun (EU) with EU API endpoint", async () => {
+    await sendWithProvider("mailgun-eu", plainMsg);
+
+    expectMailgunRequest("https://api.eu.mailgun.net/v3/example.com/messages");
+  });
+
+  test("sends via Mailgun without h:Reply-To when not provided", async () => {
+    await sendWithProvider("mailgun-us", {
+      html: "<p>Hi</p>",
+      subject: "Test",
+      text: "Hi",
+      to: validEmail("user@test.com"),
+    });
+
+    expect(fetch.getFetchFormBody().get("h:Reply-To")).toBeNull();
+  });
+
+  test("returns status code on non-OK response", async () => {
+    fetch.restubFetch(() =>
+      Promise.resolve(new Response("Error", { status: 500 })),
     );
-    expect(data.listingDate).toBe("2026-07-01T19:00:00Z");
-    expect(data.listingLocation).toBe("Town Hall");
+
+    await sendEmailExpectingError(
+      testConfig,
+      minimalMsg,
+      500,
+      "provider=resend status=500",
+    );
   });
 
-  test("sets purchaseOnly from listing flag", () => {
-    const data = buildSvgTicketData(makeEntry({ purchase_only: true }), "GBP");
-    expect(data.purchaseOnly).toBe(true);
+  test("returns undefined on fetch failure", async () => {
+    fetch.restubFetch(() => Promise.reject(new Error("Network error")));
+
+    await sendEmailExpectingError(
+      testConfig,
+      minimalMsg,
+      undefined,
+      "Network error",
+    );
   });
 
-  test("passes currency through to ticket data", () => {
-    const data = buildSvgTicketData(makeEntry(), "USD");
-    expect(data.currency).toBe("USD");
+  test("returns undefined for non-Error thrown values", async () => {
+    fetch.restubFetch(() => Promise.reject("string error"));
+
+    await sendEmailExpectingError(
+      testConfig,
+      minimalMsg,
+      undefined,
+      "string error",
+    );
+  });
+
+  test("registration delivery returns network failures", async () => {
+    fetch.restubFetch(() => Promise.reject(new TypeError("Network error")));
+
+    expect(await deliverRegistrationEmail(testConfig, minimalMsg)).toEqual({
+      delivered: false,
+      detail: "Network error",
+      status: undefined,
+    });
+  });
+
+  test("registration delivery throws internal failures", async () => {
+    const failure = new Error("Internal error");
+    fetch.restubFetch(() => Promise.reject(failure));
+
+    await expect(deliverRegistrationEmail(testConfig, minimalMsg)).rejects.toBe(
+      failure,
+    );
   });
 });
 
-describe("buildTicketAttachments", () => {
-  test("generates one attachment per entry", async () => {
-    const entries = [
-      makeEntry({}, { ticket_token: "tok1" }),
-      makeEntry({}, { ticket_token: "tok2" }),
-    ];
-    const attachments = await buildTicketAttachments(
-      soloGroups(entries),
-      "GBP",
+test("email providers have display labels", () => {
+  expect(EMAIL_PROVIDER_LABELS).toEqual({
+    "mailgun-eu": "Mailgun (EU)",
+    "mailgun-us": "Mailgun (US)",
+    postmark: "Postmark",
+    resend: "Resend",
+    sendgrid: "SendGrid",
+  });
+});
+
+describe("sendTestEmail", () => {
+  const fetch = useFetchStub();
+
+  test("sends the translated test message", async () => {
+    const status = await sendTestEmail(
+      testConfig,
+      validEmail("admin@test.com"),
     );
 
-    expect(attachments.length).toBe(2);
-    expect(attachments[0]!.filename).toBe("ticket-1.svg");
-    expect(attachments[1]!.filename).toBe("ticket-2.svg");
-    expect(attachments[0]!.contentType).toBe("image/svg+xml");
+    expect(status).toBe(200);
+    expect(fetch.callCount()).toBe(1);
+    const body = fetch.getFetchJsonBody();
+    expect(body.to).toEqual(["admin@test.com"]);
+    expect(body.html).toBe(`<p>${t("fields.email.test_body")}</p>`);
+    expect(body.subject).toBe(t("fields.email.test_subject"));
+    expect(body.text).toBe(t("fields.email.test_body"));
+  });
+});
+
+describe("sendEmail with attachments", () => {
+  const fetch = useFetchStub();
+
+  const attachment = {
+    content: btoa("<svg>test</svg>"),
+    contentType: "image/svg+xml",
+    filename: "ticket.svg",
+  };
+  const msgWithAttachment: EmailMessage = {
+    attachments: [attachment],
+    html: "<p>Hi</p>",
+    subject: "Tickets",
+    text: "Hi",
+    to: validEmail("user@test.com"),
+  };
+
+  test("Resend includes attachments with filename and content", async () => {
+    await sendEmail(testConfig, msgWithAttachment);
+
+    expect(fetch.getFetchJsonBody().attachments).toEqual([
+      { content: attachment.content, filename: "ticket.svg" },
+    ]);
   });
 
-  test("uses 'ticket.svg' filename for single entry", async () => {
-    const attachments = await buildTicketAttachments(
-      soloGroups([makeEntry()]),
-      "GBP",
-    );
+  test("Postmark includes Attachments with Name, Content, ContentType", async () => {
+    await sendWithProvider("postmark", msgWithAttachment);
 
-    expect(attachments.length).toBe(1);
-    expect(attachments[0]!.filename).toBe("ticket.svg");
+    expect(fetch.getFetchJsonBody().Attachments).toEqual([
+      {
+        Content: attachment.content,
+        ContentType: "image/svg+xml",
+        Name: "ticket.svg",
+      },
+    ]);
   });
 
-  test("attachment content is base64-encoded UTF-8 SVG", async () => {
-    const attachments = await buildTicketAttachments(
-      soloGroups([makeEntry()]),
-      "GBP",
-    );
+  test("SendGrid includes attachments with content, filename, type, disposition", async () => {
+    await sendWithProvider("sendgrid", msgWithAttachment);
 
-    const decoded = decodeAttachmentContent(attachments[0]!);
-    expect(decoded).toContain("<?xml");
-    expect(decoded).toContain("<svg");
-    expect(decoded).toContain("</svg>");
+    expect(fetch.getFetchJsonBody().attachments).toEqual([
+      {
+        content: attachment.content,
+        disposition: "attachment",
+        filename: "ticket.svg",
+        type: "image/svg+xml",
+      },
+    ]);
   });
 
-  test("attachment preserves non-ASCII characters via UTF-8 encoding", async () => {
-    const attachments = await buildTicketAttachments(
-      soloGroups([makeEntry({ location: "Zurich", name: "Cafe Musik" })]),
-      "GBP",
-    );
+  test("Mailgun appends attachment as Blob to FormData", async () => {
+    await sendWithProvider("mailgun-us", msgWithAttachment);
 
-    const decoded = decodeAttachmentContent(attachments[0]!);
-    expect(decoded).toContain("Cafe Musik");
-    expect(decoded).toContain("Zurich");
+    const body = fetch.getFetchFormBody();
+    const file = body.get("attachment") as File;
+    expect(file).toBeInstanceOf(File);
+    expect(file.name).toBe("ticket.svg");
+    expect(file.type).toBe("image/svg+xml");
   });
 
-  test("collapses a hidden package into one SVG showing the package name, not members", async () => {
-    const entries = [
-      makeEntry(
-        { name: "Secret Member A" },
-        { price_paid: "1000", quantity: 2, ticket_token: "pkgtok" },
-      ),
-      makeEntry(
-        { name: "Secret Member B" },
-        { price_paid: "500", quantity: 1, ticket_token: "pkgtok" },
-      ),
-    ];
-    const attachments = await buildTicketAttachments(
-      [{ entries, hiddenPackageName: "Weekend Bundle" }],
-      "GBP",
-    );
+  test("omits attachments field when no attachments provided", async () => {
+    await sendEmail(testConfig, minimalMsg);
 
-    expect(attachments.length).toBe(1);
-    expect(attachments[0]!.filename).toBe("ticket.svg");
-    const decoded = decodeAttachmentContent(attachments[0]!);
-    expect(decoded).toContain("Weekend Bundle");
-    expect(decoded).not.toContain("Secret Member A");
-    expect(decoded).not.toContain("Secret Member B");
-    // The bundle's summed quantity heads the single ticket (2 + 1).
-    expect(decoded).toContain("Qty: 3");
-  });
-
-  test("a dated hidden package's collapsed SVG keeps the booked date", async () => {
-    const entries = [
-      makeEntry(
-        { name: "Secret Cabin" },
-        {
-          date: "2026-08-01",
-          end_date: "2026-08-03",
-          price_paid: "1000",
-          quantity: 1,
-          ticket_token: "datedtok",
-        },
-      ),
-    ];
-    const attachments = await buildTicketAttachments(
-      [{ entries, hiddenPackageName: "Dated Bundle" }],
-      "GBP",
-    );
-    const decoded = decodeAttachmentContent(attachments[0]!);
-    expect(decoded).toContain("Dated Bundle");
-    expect(decoded).not.toContain("Secret Cabin");
-    // Hiding the members must not lose the buyer's booked date.
-    expect(decoded).toContain("1 August 2026");
-  });
-
-  test("a mixed order conceals its hidden bundle beside a normal ticket", async () => {
-    const attachments = await buildTicketAttachments(
-      [
-        {
-          entries: [makeEntry({ name: "Secret Member" }, { quantity: 2 })],
-          hiddenPackageName: "Mystery Box",
-        },
-        { entries: [makeEntry({ name: "Lantern" })] },
-      ],
-      "GBP",
-    );
-
-    expect(attachments.length).toBe(2);
-    const first = decodeAttachmentContent(attachments[0]!);
-    const second = decodeAttachmentContent(attachments[1]!);
-    expect(first).toContain("Mystery Box");
-    expect(first).not.toContain("Secret Member");
-    expect(second).toContain("Lantern");
+    expect(fetch.getFetchJsonBody().attachments).toBeUndefined();
   });
 });
